% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AR6regions}
\alias{AR6regions}
\title{IPCC AR6 regions}
\format{
An object of class \code{list} of length 62.
}
\source{
\url{https://github.com/IPCC-WG1/Atlas/blob/main/reference-regions/IPCC-WGI-reference-regions-v4_coordinates.csv}
}
\usage{
AR6regions
}
\description{
Boundaries of the IPCC AR6 regions
}
\keyword{datasets}
