% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReleasedTools.R
\name{countGithubDownloads}
\alias{countGithubDownloads}
\title{Github downloads}
\usage{
countGithubDownloads(url)
}
\arguments{
\item{url}{string, Github API URL to retrieve the desired releases.
Should look something like 'https://api.github.com/repos/repo-name/releases'}
}
\value{
A data frame containing the downloads count for all assets from all releases found in the provided URL.
}
\description{
Count the number of downloads from GitHub.
}
\examples{
DF=countGithubDownloads('https://api.github.com/repos/BaRatin-tools/BaRatinAGE/releases')
plot(DF$published,DF$downloads)
}
