% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{dailyToMonthly}
\alias{dailyToMonthly}
\title{Daily to monthly.}
\usage{
dailyToMonthly(x, stat = mean, ...)
}
\arguments{
\item{x}{data frame, should contain at least three columns named 'day', 'month' and 'value'.}

\item{stat}{function, statistics to be applied.}

\item{...}{other arguments passed to function stat.}
}
\value{
A data frame with the following columns:
\enumerate{
  \item year
  \item month
  \item value
}
}
\description{
Transform a daily time series into a monthly one.
}
\examples{
DF=cbind(getCalendar(2020),value=rnorm(366))
dailyToMonthly(DF)
dailyToMonthly(DF,stat=max)
dailyToMonthly(DF,stat=quantile,probs=0.75)
}
