% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{daysInMonth}
\alias{daysInMonth}
\title{Number of days in a calendar month}
\usage{
daysInMonth(months = 1:12, isLeap = FALSE)
}
\arguments{
\item{months}{integer vector, months (between 1 and 12).}

\item{isLeap}{logical, if TRUE, all Februaries in vector months will have 29 days.}
}
\value{
A vector with same size as months containing the corresponding number of days.
}
\description{
Return the number of days in given calendar months
}
\examples{
daysInMonth()
}
