% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderAndFiles.R
\name{dirDiff}
\alias{dirDiff}
\title{Directory Diff}
\usage{
dirDiff(master.dir, copy.dir, hours = 24, fileOnly = TRUE)
}
\arguments{
\item{master.dir}{character string, master directory}

\item{copy.dir}{character string, copy directory}

\item{hours}{numeric, modificaton time span}

\item{fileOnly}{logical, list only files or subfolders as well?}
}
\value{
A list with the following fields:
\enumerate{
  \item new, list of new files (in master but not in copy)
  \item removed, list of removed files (in copy but not in master)
  \item sizeChanged, list of size-changed files (in both copy and master but with different sizes)
  \item modified, list of modified files (according to OS-specific 'time of last modif', unclear what this does exactly...)
}
}
\description{
Get a diff between a master and a copy directories, i.e. list files
and folders that have been added, removed, sized-changed or
os-modified in the last X hours.
}
\examples{
D0=tempdir()
dir.create(file.path(D0,'D1'))
file.create(file.path(D0,'D1','f1.txt'))
file.create(file.path(D0,'D1','f2.txt'))
file.create(file.path(D0,'D1','f3.txt'))
dir.create(file.path(D0,'D2'))
file.create(file.path(D0,'D2','f1.txt'))
file.create(file.path(D0,'D2','f4.txt'))
dirDiff(file.path(D0,'D1'),file.path(D0,'D2'))
# Cleaning up: delete D1 and D2
unlink(c(file.path(D0,'D1'),file.path(D0,'D2')))
}
