% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{downloadHydroPortailData}
\alias{downloadHydroPortailData}
\title{HydroPortail batch downloader}
\usage{
downloadHydroPortailData(
  saveTo,
  what = c("stations", "sites"),
  url_prefix = "https://bnum.din.gouv.fr/mdrive/index.php/s/",
  url_id = c("AcdANJytrNy7Hrs", "qCcoDTJ9gaALQc2", "Y2oAEKmf7BPExyR", "RL4eqXnmL3ciZGy",
    "J79dFNWbbZFi6CK", "JABp59e9tTySeRR", "SaT3wXMkMnCwwrd"),
  url_fsuffix = c("_09", "_AE", "_FJ", "_PT", "_UZ", "_KN", "_O")
)
}
\arguments{
\item{saveTo}{character string, destination folder}

\item{what}{character string, stations or sites ?}

\item{url_prefix}{character string, prefix of the download URL}

\item{url_id}{character string, id of the download URL}

\item{url_fsuffix}{character string, suffix of the downloaded files (without extension)}
}
\value{
nothing - just write files to disk.
}
\description{
Download HydroPortail data from
https://data.ofb.fr/catalogue/data-eaufrance/eng/catalog.search#/metadata/cae2a6c7-4429-4c89-9da2-44ada7735520
}
\examples{
dest=tempdir()
downloadHydroPortailData(saveTo=dest,url_id='',url_fsuffix='_09')
}
