% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{extractQJfilesFromHP}
\alias{extractQJfilesFromHP}
\title{HydroPortail QJ getter}
\usage{
extractQJfilesFromHP(tarDir, saveTo = tarDir, subfolder = "QJ")
}
\arguments{
\item{tarDir}{character string, directory containing the HydroPortail tar files}

\item{saveTo}{character string, destination folder where QJ files will be saved}

\item{subfolder}{character string, subfolder of saveTo where QJ files will be saved}
}
\value{
nothing - just write files to disk.
}
\description{
Extract daily streamflow files from the tar files dowloaded with downloadData_HP
}
\examples{
dest=tempdir()
downloadData_HP(saveTo=dest,url_id='',url_fsuffix='_09')
extractQJfilesFromHP(tarDir=dest)
}
