% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maps.R
\name{getAspectRatio}
\alias{getAspectRatio}
\title{Compute y/x Aspect Ratio}
\usage{
getAspectRatio(lat, ...)
}
\arguments{
\item{lat}{numeric, latitude for which the ratio is sought.}

\item{...}{other arguments passed to function haversine.}
}
\value{
a numeric giving the y/x aspect ratio
}
\description{
Compute the y/x aspect ratio for a map in lon-lat, centered around a given latitude.
The ratio is computed as the ratio of distances traveled with a 1 degree move equator-ward and eastward.
}
\examples{
getAspectRatio(0)
getAspectRatio(50)
getAspectRatio(-50)
}
