% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maps.R
\name{getBBox}
\alias{getBBox}
\title{Get bounding box}
\usage{
getBBox(bbox = "World")
}
\arguments{
\item{region}{character, bounding box as a geographical description.
    Available:
\enumerate{
  \item 'All','World' (Antartica removed for the latter)
  \item Continents: 'Africa','America','Asia', Europe', 'Oceania'.
  \item Continental regions: 'SouthAmerica', 'CentralAmerica',
      'NorthAmerica', 'WesternEurope', 'Mediterranean'.
  \item Oceans: 'Atlantic','Indian','Pacific'.
  \item Oceanic regions: 'NorthAtlantic','SouthAtlantic','IndianPacific'.
  \item Countries: 'Australia','NewZealand','France'.
}}
}
\value{
a vector of size 4 (minlon, maxlon, minlat, maxlat), or NULL
    if the region is not available.
}
\description{
Get bounding box from a geographical description
}
\examples{
getBBox('SouthAtlantic')
getBBox('Europe')
getBBox('Mediterranean')
getBBox('NewZealand')
}
