% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{getCalendar}
\alias{getCalendar}
\title{Get calendar}
\usage{
getCalendar(years)
}
\arguments{
\item{years}{integer.}
}
\value{
A data frame with the following columns:
\enumerate{
  \item day
  \item month
  \item year
  \item weekday
}
}
\description{
Get calendar for given civil year(s)
}
\examples{
getCalendar(2020)
}
