% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{getHydroYear}
\alias{getHydroYear}
\title{Hydrological Year}
\usage{
getHydroYear(months, years, hydroYear = 1:12)
}
\arguments{
\item{months}{integer vector, months (between 1 and 12).}

\item{years}{integer vector, same size as months, years.}

\item{hydroYear}{integer vector, hydro-year definition.}
}
\value{
A vector with same size as years containing the corresponding hydro-year.
}
\description{
Compute hydrological (or meteorological or whatever) year
}
\examples{
getHydroYear(months=1:12,years=rep(2020,12),hydroYear=c(9,10,11))
getHydroYear(months=1:12,years=rep(2020,12),hydroYear=c(12,1,2))
}
