% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{getMonthlyStat}
\alias{getMonthlyStat}
\title{Monthly stats}
\usage{
getMonthlyStat(x, stat = mean, ...)
}
\arguments{
\item{x}{data frame, should contain at least two columns named 'month' and 'value'.}

\item{stat}{function, statistics to be applied.}

\item{...}{other arguments passed to function stat.}
}
\value{
A vector of size 12 containing the monthly statistics
}
\description{
Compute interannual monthly statistics from a time series
}
\examples{
data(nottem) # 	Average Monthly Temperatures (F) at Nottingham, 1920-1939
DF=data.frame(year=rep(1920:1939,each=12),month=1:12,value=nottem)
getMonthlyStat(DF)
getMonthlyStat(DF,stat=quantile,probs=0.75)
}
