% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalProcessing.R
\name{getPeriodogram}
\alias{getPeriodogram}
\title{Periodogram}
\usage{
getPeriodogram(x)
}
\arguments{
\item{x}{numeric vector, data, assumed to represent daily values.
WARNING: the time step is assumed to be daily and constant across the whole series.
This implies in particular that time steps with missing values should not be removed,
but should rather be assigned a NA value. NA values will be internally replaced with
the mean of x, which is probably a poor approach when many values are missing.}
}
\value{
A periodogram object, which is just a data frame with columns: frequency, period_days, period_years and power.
}
\description{
Compute a periodogram from a daily time series
}
\examples{
# 40 years of time steps
x=1:(365.25*40)
# Mix of 3 frequencies
y=sin(2*pi*x/365.25)+0.7*sin(2*pi*x/(0.5*365.25))+0.5*sin(2*pi*x/(6*365.25))
plot(x,y,type='l')
# Get periodogram and plot it
per=getPeriodogram(y)
plot(per$period_year,per$power,log='x',type='l')
}
