% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getQRcode}
\alias{getQRcode}
\title{QR code}
\usage{
getQRcode(url, colorBkg = "white", colorTile = "black", alpha = 1)
}
\arguments{
\item{url}{character string, URL}

\item{colorBkg}{color, background color}

\item{colorTile}{color, tile color}
}
\value{
A ggplot
}
\description{
Get QRcode associated with a URL as a ggplot
}
\examples{
getQRcode('https://globxblog.github.io/')
}
