% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maps.R
\name{getWorldMap}
\alias{getWorldMap}
\title{World map}
\usage{
getWorldMap(
  fill = "gray20",
  borderColor = fill,
  borderWidth = 0.1,
  region = "World",
  theme = theme_void(),
  addFrame = T
)
}
\arguments{
\item{fill}{fill color.}

\item{borderColor}{border color.}

\item{borderWidth}{numeric, border width.}

\item{region}{character, bounding box as a geographical description}

\item{theme}{ggplot theme}

\item{addFrame}{logical, add a frame (mostly useful when theme=theme_void())}
}
\value{
a ggplot object
}
\description{
Create a World Map as a ggplot object
}
\examples{
getWorldMap(region='SouthAtlantic')
getWorldMap(region='Europe')
getWorldMap(region='Mediterranean',borderColor='gray50')
getWorldMap(region='NewZealand')
}
