% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maps.R
\name{haversine}
\alias{haversine}
\title{Haversine distance}
\usage{
haversine(pt1, pt2, R = 6371)
}
\arguments{
\item{pt1, pt2}{numeric vectors of size at least 2 starting with (lon,lat)
coordinates. If a third coordinate is present (e.g. elevation), it
will be ignored.}

\item{R}{numeric, Earth's radius in km.}
}
\value{
a numeric giving the distance between the two points in km
}
\description{
Compute the haversine distance between two (lon,lat) points.
Uses Salvador's script in this page:
\url{https://stackoverflow.com/questions/27928/calculate-distance-between-two-latitude-longitude-points-haversine-formula}
}
\examples{
pt1=c(-32.917,151.75) #Newcastle, AU
pt2=c(43.2964,5.37) #Marseille, FR
haversine(pt1,pt2)
}
