% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{isLeapYear}
\alias{isLeapYear}
\title{Leap Year}
\usage{
isLeapYear(years)
}
\arguments{
\item{years}{integer vector.}
}
\value{
A logical vector with same size as years.
}
\description{
Are these leap years?
}
\examples{
isLeapYear(c(1900,2000,2010,2011,2012))
}
