% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{monthName}
\alias{monthName}
\title{Month name}
\usage{
monthName(months = 1:12, nchar = NULL)
}
\arguments{
\item{months}{integer vector, months (between 1 and 12).}

\item{nchar}{positive integer, month names are trimmed to nchar characters.}
}
\value{
A character vector with same size as months containing the corresponding names.
}
\description{
Return the name of calendar months
}
\examples{
monthName()
monthName(nchar=1)
monthName(nchar=3)
}
