% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalProcessing.R
\name{plot.periodogram}
\alias{plot.periodogram}
\title{Plot Periodogram}
\usage{
\method{plot}{periodogram}(x, verticalLines = c(0.5, 1, 5, 10), ...)
}
\arguments{
\item{x}{periodogram object, periodogram computed by function [getPeriodogram()].}

\item{verticalLines}{numeric vector, periods (in years) where vertical lines are drawn}

\item{...}{arguments passed to other methods}
}
\value{
a ggplot of the periodogram.
}
\description{
Plot a periodogram resulting from a call to function getPeriodogram
}
\examples{
# 40 years of time steps
x=1:(365.25*40)
# Mix of 3 frequencies
y=sin(2*pi*x/365.25)+0.7*sin(2*pi*x/(0.5*365.25))+0.5*sin(2*pi*x/(6*365.25))
plot(x,y,type='l')
# Get periodogram and plot it
per=getPeriodogram(y)
plot(per,verticalLines=c(0.5,1,6))
}
