% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAndWrite.R
\name{read_OzRefNetwork_M}
\alias{read_OzRefNetwork_M}
\title{Read Oz monthly data}
\usage{
read_OzRefNetwork_M(f)
}
\arguments{
\item{f}{character string, file.}
}
\value{
A data frame with the following columns:
\enumerate{
  \item year
  \item month
  \item value
}
}
\description{
Read monthly data in the format used by the BoM for
the Australian Climate Observation Reference Network
}
\examples{
#
}
