% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAndWrite.R
\name{read_OzRefNetwork_Q_D}
\alias{read_OzRefNetwork_Q_D}
\title{Read Oz RHN data}
\usage{
read_OzRefNetwork_Q_D(f)
}
\arguments{
\item{f}{character string, file.}
}
\value{
A data frame with the following columns:
\enumerate{
  \item year
  \item month
  \item day
  \item value
  \item qcode (quality code)
}
}
\description{
Read daily data in the format used by the BoM for
the Australian Hydrometric Reference Network
}
\examples{
#
}
