% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAndWrite.R
\name{read_OzRefNetwork_T_D}
\alias{read_OzRefNetwork_T_D}
\title{Read Oz ACORN-SAT v2 temperature data}
\usage{
read_OzRefNetwork_T_D(f)
}
\arguments{
\item{f}{character string, file.}
}
\value{
A data frame with the following columns:
\enumerate{
  \item year
  \item month
  \item day
  \item value
}
}
\description{
Read daily data in the format used by the BoM for
the Australian CLimate Observation Reference Network
}
\examples{
#
}
