% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAndWrite.R
\name{read_tabular_M}
\alias{read_tabular_M}
\title{Read tabular data}
\usage{
read_tabular_M(f, ...)
}
\arguments{
\item{f}{character string, file.}

\item{...}{other arguments passed to function read.table}
}
\value{
A data frame with the following columns:
\enumerate{
  \item year
  \item month
  \item value
}
}
\description{
Read monthly data in a tabular format (year then 12 columns).
Only the first 13 columns are considered, anything on the right of the
13th (e.g. an annual value in some datasets) is ignored.
}
\examples{
#
}
