% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SREXregions}
\alias{SREXregions}
\title{IPCC SREX regions}
\format{
An object of class \code{list} of length 33.
}
\source{
\url{https://www.ipcc-data.org/guidelines/pages/ar5_regions.html}
}
\usage{
SREXregions
}
\description{
Boundaries of the SREX regions used in various IPCC reports (AR5 in particular)
}
\keyword{datasets}
